Organization and fixes
Re-structured to store embeds, modals, logger all in the same file to make access to re-used code easily for future updates. I also finalized the queue system and worked out registration and stats commands to dynamically generate based on the data avaliable. Planning next to work on the draft class, and make it well optimized.